HOP-2477 Move interface IValueMetaConverter to package org.apache.hop.core.row
